% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4RuBPCarboxyPEPCarboxy.R
\name{C4RuBPCarboxyPEPCarboxy}
\alias{C4RuBPCarboxyPEPCarboxy}
\title{Calculate C4 RuBP carboxylation and PEP carboxylation photosynthesis rate}
\usage{
C4RuBPCarboxyPEPCarboxy(Vcmax25, J25, Ca, Tleaf, Phis, VPD, StartV)
}
\arguments{
\item{Vcmax25}{Maximal velocity of Rubisco carboxylation at 25°C (μmol m-2 s-1).}

\item{J25}{Electron transport at a specific light intensity at 25°C. The value should be adjusted at different light intensities (μmol m-2 s-1).}

\item{Ca}{CO2 concentration in the atmosphere (ppm).}

\item{Tleaf}{Leaf temperature (oC).}

\item{Phis}{Soil water potential (Mpa).}

\item{VPD}{Vapour pressure deficit (kPa).}

\item{StartV}{A vector that gives the start points for the estimation (c(A,rs,fl)).A is photosynthesis rate (μmol m-2 s-1),rs is stomatal resistance (μmol-1 m2 s1, the reciprocal of stomatal conductance) and fl is the energy allocation ratio of leaf to the total plant.}
}
\value{
This package will return a dataframe of the numerical solution of the model using multiroot in the package "rootSolve".

Model$root: A vector or solution of c(A,rs,fl).
}
\description{
Calculate C4 RuBP carboxylation and PEP carboxylation photosynthesis rate using von Caemmerer photosynthesis model for C4 pathway (von Caemmerer S, 2000) with optimality to stomatal resistance and energy allocation to leaf. The use of the function is as follows: C4RuBPCarboxyPEPCarboxy(Vcmax25,J25,Ca,Tleaf,Phis,VPD,StartV). The package "rootSolve" should be installed before using the function.
}
